requirements.txt
streamlit==1.5.0
tensorflow==2.12.0
numpy
pillow
